Build insert and find functions
Write function to insert values into the tree and ensure the tree
is rebalanced after the insertion.  Write find function to use within
the insert function as a helper to confirm if the value already
exists in the tree.  If the value exists do not insert into the tree,
else insert the value in the tree.